+++ src/app/pages/user-profile/user-profile.component.scss
@@ -37,8 +37,6 @@ mat-accordion {
   }
 }
 
-// All addresses cards
-
 mat-card-content {
   display: flex;
   justify-content: space-between;
